Instrument 5.1+ instead
