+++ instrumentation/lettuce/lettuce-5.1/src/test/groovy/io/opentelemetry/auto/instrumentation/lettuce/v5_1/LettuceAsyncClientTest.groovy
@@ -13,22 +13,33 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.opentelemetry.auto.instrumentation.lettuce.v5_2
+package io.opentelemetry.auto.instrumentation.lettuce.v5_1
 
 import io.lettuce.core.ClientOptions
-
+import io.lettuce.core.ConnectionFuture
 import io.lettuce.core.RedisClient
-import io.lettuce.core.RedisConnectionException
+import io.lettuce.core.RedisFuture
+import io.lettuce.core.RedisURI
 import io.lettuce.core.api.StatefulConnection
+import io.lettuce.core.api.async.RedisAsyncCommands
 import io.lettuce.core.api.sync.RedisCommands
+import io.lettuce.core.codec.StringCodec
 import io.opentelemetry.auto.test.AgentTestRunner
 import io.opentelemetry.auto.test.utils.PortUtils
 import redis.embedded.RedisServer
 import spock.lang.Shared
+import spock.util.concurrent.AsyncConditions
+
+import java.util.concurrent.ExecutionException
+import java.util.concurrent.TimeUnit
+import java.util.function.BiConsumer
+import java.util.function.BiFunction
+import java.util.function.Consumer
+import java.util.function.Function
 
 import static io.opentelemetry.trace.Span.Kind.CLIENT
 
-class LettuceSyncClientTest extends AgentTestRunner {
+class LettuceAsyncClientTest extends AgentTestRunner {
   public static final String HOST = "127.0.0.1"
   public static final int DB_INDEX = 0
   // Disable autoreconnect so we do not get stray traces popping up on server shutdown
@@ -46,8 +57,6 @@ class LettuceSyncClientTest extends AgentTestRunner {
   String dbUriNonExistent
   @Shared
   String embeddedDbUri
-  @Shared
-  String embeddedDbLocalhostUri
 
   @Shared
   RedisServer redisServer
@@ -61,6 +70,7 @@ class LettuceSyncClientTest extends AgentTestRunner {
 
   RedisClient redisClient
   StatefulConnection connection
+  RedisAsyncCommands<String, ?> asyncCommands
   RedisCommands<String, ?> syncCommands
 
   def setupSpec() {
@@ -70,7 +80,6 @@ class LettuceSyncClientTest extends AgentTestRunner {
     dbAddrNonExistent = HOST + ":" + incorrectPort + "/" + DB_INDEX
     dbUriNonExistent = "redis://" + dbAddrNonExistent
     embeddedDbUri = "redis://" + dbAddr
-    embeddedDbLocalhostUri = "redis://localhost:" + port + "/" + DB_INDEX
 
     redisServer = RedisServer.builder()
     // bind to localhost to avoid firewall popup
@@ -83,15 +92,18 @@ class LettuceSyncClientTest extends AgentTestRunner {
   def setup() {
     redisClient = RedisClient.create(embeddedDbUri)
 
+    println "Using redis: $redisServer.args"
     redisServer.start()
+    redisClient.setOptions(CLIENT_OPTIONS)
+
     connection = redisClient.connect()
+    asyncCommands = connection.async()
     syncCommands = connection.sync()
 
     syncCommands.set("TESTKEY", "TESTVAL")
-    syncCommands.hmset("TESTHM", testHashMap)
 
-    // 2 sets
-    TEST_WRITER.waitForTraces(2)
+    // 1 set
+    TEST_WRITER.waitForTraces(1)
     TEST_WRITER.clear()
   }
 
@@ -100,15 +112,18 @@ class LettuceSyncClientTest extends AgentTestRunner {
     redisServer.stop()
   }
 
-  def "connect"() {
+  def "connect using get on ConnectionFuture"() {
     setup:
     RedisClient testConnectionClient = RedisClient.create(embeddedDbUri)
     testConnectionClient.setOptions(CLIENT_OPTIONS)
 
     when:
-    StatefulConnection connection = testConnectionClient.connect()
+    ConnectionFuture connectionFuture = testConnectionClient.connectAsync(StringCodec.UTF8,
+      new RedisURI(HOST, port, 3, TimeUnit.SECONDS))
+    StatefulConnection connection = connectionFuture.get()
 
     then:
+    connection != null
     // Lettuce tracing does not trace connect
     assertTraces(0) {}
 
@@ -116,23 +131,27 @@ class LettuceSyncClientTest extends AgentTestRunner {
     connection.close()
   }
 
-  def "connect exception"() {
+  def "connect exception inside the connection future"() {
     setup:
     RedisClient testConnectionClient = RedisClient.create(dbUriNonExistent)
     testConnectionClient.setOptions(CLIENT_OPTIONS)
 
     when:
-    testConnectionClient.connect()
+    ConnectionFuture connectionFuture = testConnectionClient.connectAsync(StringCodec.UTF8,
+      new RedisURI(HOST, incorrectPort, 3, TimeUnit.SECONDS))
+    StatefulConnection connection = connectionFuture.get()
 
     then:
-    thrown RedisConnectionException
+    connection == null
+    thrown ExecutionException
     // Lettuce tracing does not trace connect
     assertTraces(0) {}
   }
 
-  def "set command"() {
+  def "set command using Future get with timeout"() {
     setup:
-    String res = syncCommands.set("TESTSETKEY", "TESTSETVAL")
+    RedisFuture<String> redisFuture = asyncCommands.set("TESTSETKEY", "TESTSETVAL")
+    String res = redisFuture.get(3, TimeUnit.SECONDS)
 
     expect:
     res == "OK"
@@ -160,47 +179,25 @@ class LettuceSyncClientTest extends AgentTestRunner {
       }
     }
   }
-  def "set command localhost"() {
-    setup:
-    RedisClient testConnectionClient = RedisClient.create(embeddedDbLocalhostUri)
-    testConnectionClient.setOptions(CLIENT_OPTIONS)
-    StatefulConnection connection = testConnectionClient.connect()
-    String res = connection.sync().set("TESTSETKEY", "TESTSETVAL")
 
-    expect:
-    res == "OK"
-    assertTraces(1) {
-      trace(0, 1) {
-        span(0) {
-          operationName "SET"
-          spanKind CLIENT
-          errored false
-          tags {
-            "net.transport" "IP.TCP"
-            "net.peer.ip" "127.0.0.1"
-            "net.peer.name" "localhost"
-            "net.peer.port" port
-            "db.url" "redis://localhost:$port"
-            "db.type" "redis"
-            "db.statement" "key<TESTSETKEY> value<TESTSETVAL>"
-          }
-          event(0) {
-            eventName "redis.encode.start"
-          }
-          event(1) {
-            eventName "redis.encode.end"
-          }
+  def "get command chained with thenAccept"() {
+    setup:
+    def conds = new AsyncConditions()
+    Consumer<String> consumer = new Consumer<String>() {
+      @Override
+      void accept(String res) {
+        conds.evaluate {
+          assert res == "TESTVAL"
         }
       }
     }
-  }
 
-  def "get command"() {
-    setup:
-    String res = syncCommands.get("TESTKEY")
+    when:
+    RedisFuture<String> redisFuture = asyncCommands.get("TESTKEY")
+    redisFuture.thenAccept(consumer)
 
-    expect:
-    res == "TESTVAL"
+    then:
+    conds.await()
     assertTraces(1) {
       trace(0, 1) {
         span(0) {
@@ -226,12 +223,38 @@ class LettuceSyncClientTest extends AgentTestRunner {
     }
   }
 
-  def "get non existent key command"() {
+  // to make sure instrumentation's chained completion stages won't interfere with user's, while still
+  // recording metrics
+  def "get non existent key command with handleAsync and chained with thenApply"() {
     setup:
-    String res = syncCommands.get("NON_EXISTENT_KEY")
+    def conds = new AsyncConditions()
+    final String successStr = "KEY MISSING"
+    BiFunction<String, Throwable, String> firstStage = new BiFunction<String, Throwable, String>() {
+      @Override
+      String apply(String res, Throwable throwable) {
+        conds.evaluate {
+          assert res == null
+          assert throwable == null
+        }
+        return (res == null ? successStr : res)
+      }
+    }
+    Function<String, Object> secondStage = new Function<String, Object>() {
+      @Override
+      Object apply(String input) {
+        conds.evaluate {
+          assert input == successStr
+        }
+        return null
+      }
+    }
 
-    expect:
-    res == null
+    when:
+    RedisFuture<String> redisFuture = asyncCommands.get("NON_EXISTENT_KEY")
+    redisFuture.handleAsync(firstStage).thenApply(secondStage)
+
+    then:
+    conds.await()
     assertTraces(1) {
       trace(0, 1) {
         span(0) {
@@ -257,12 +280,24 @@ class LettuceSyncClientTest extends AgentTestRunner {
     }
   }
 
-  def "command with no arguments"() {
+  def "command with no arguments using a biconsumer"() {
     setup:
-    def keyRetrieved = syncCommands.randomkey()
+    def conds = new AsyncConditions()
+    BiConsumer<String, Throwable> biConsumer = new BiConsumer<String, Throwable>() {
+      @Override
+      void accept(String keyRetrieved, Throwable throwable) {
+        conds.evaluate {
+          assert keyRetrieved != null
+        }
+      }
+    }
 
-    expect:
-    keyRetrieved != null
+    when:
+    RedisFuture<String> redisFuture = asyncCommands.randomkey()
+    redisFuture.whenCompleteAsync(biConsumer)
+
+    then:
+    conds.await()
     assertTraces(1) {
       trace(0, 1) {
         span(0) {
@@ -287,44 +322,43 @@ class LettuceSyncClientTest extends AgentTestRunner {
     }
   }
 
-  def "list command"() {
+  def "hash set and then nest apply to hash getall"() {
     setup:
-    long res = syncCommands.lpush("TESTLIST", "TESTLIST ELEMENT")
+    def conds = new AsyncConditions()
 
-    expect:
-    res == 1
-    assertTraces(1) {
-      trace(0, 1) {
-        span(0) {
-          operationName "LPUSH"
-          spanKind CLIENT
-          errored false
-          tags {
-            "net.transport" "IP.TCP"
-            "net.peer.ip" "127.0.0.1"
-            "net.peer.port" port
-            "db.url" "redis://127.0.0.1:$port"
-            "db.type" "redis"
-            "db.statement" "key<TESTLIST> value<TESTLIST ELEMENT>"
-          }
-          event(0) {
-            eventName "redis.encode.start"
-          }
-          event(1) {
-            eventName "redis.encode.end"
-          }
+    when:
+    RedisFuture<String> hmsetFuture = asyncCommands.hmset("TESTHM", testHashMap)
+    hmsetFuture.thenApplyAsync(new Function<String, Object>() {
+      @Override
+      Object apply(String setResult) {
+        conds.evaluate {
+          assert setResult == "OK"
         }
+        RedisFuture<Map<String, String>> hmGetAllFuture = asyncCommands.hgetall("TESTHM")
+        hmGetAllFuture.exceptionally(new Function<Throwable, Map<String, String>>() {
+          @Override
+          Map<String, String> apply(Throwable throwable) {
+            println("unexpected:" + throwable.toString())
+            throwable.printStackTrace()
+            assert false
+            return null
+          }
+        })
+        hmGetAllFuture.thenAccept(new Consumer<Map<String, String>>() {
+          @Override
+          void accept(Map<String, String> hmGetAllResult) {
+            conds.evaluate {
+              assert testHashMap == hmGetAllResult
+            }
+          }
+        })
+        return null
       }
-    }
-  }
+    })
 
-  def "hash set command"() {
-    setup:
-    def res = syncCommands.hmset("user", testHashMap)
-
-    expect:
-    res == "OK"
-    assertTraces(1) {
+    then:
+    conds.await()
+    assertTraces(2) {
       trace(0, 1) {
         span(0) {
           operationName "HMSET"
@@ -336,7 +370,7 @@ class LettuceSyncClientTest extends AgentTestRunner {
             "net.peer.port" port
             "db.url" "redis://127.0.0.1:$port"
             "db.type" "redis"
-            "db.statement" "key<user> key<firstname> value<John> key<lastname> value<Doe> key<age> value<53>"
+            "db.statement" "key<TESTHM> key<firstname> value<John> key<lastname> value<Doe> key<age> value<53>"
           }
           event(0) {
             eventName "redis.encode.start"
@@ -346,17 +380,7 @@ class LettuceSyncClientTest extends AgentTestRunner {
           }
         }
       }
-    }
-  }
-
-  def "hash getall command"() {
-    setup:
-    Map<String, String> res = syncCommands.hgetall("TESTHM")
-
-    expect:
-    res == testHashMap
-    assertTraces(1) {
-      trace(0, 1) {
+      trace(1, 1) {
         span(0) {
           operationName "HGETALL"
           spanKind CLIENT
@@ -378,23 +402,5 @@ class LettuceSyncClientTest extends AgentTestRunner {
         }
       }
     }
-  }
-
-  def "debug segfault command (returns void) with no argument produces no span"() {
-    setup:
-    syncCommands.debugSegfault()
-
-    expect:
-    // lettuce tracing does not trace debug
-    assertTraces(0) {}
-  }
-
-  def "shutdown command (returns void) produces no span"() {
-    setup:
-    syncCommands.shutdown(false)
-
-    expect:
-    // lettuce tracing does not trace shutdown
-    assertTraces(0) {}
   }
 }
