+++ instrumentation/lettuce/lettuce-5.1/src/test/groovy/io/opentelemetry/auto/instrumentation/lettuce/v5_1/LettuceReactiveClientTest.groovy
@@ -13,33 +13,26 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.opentelemetry.auto.instrumentation.lettuce.v5_2
+package io.opentelemetry.auto.instrumentation.lettuce.v5_1
 
 import io.lettuce.core.ClientOptions
-import io.lettuce.core.ConnectionFuture
 import io.lettuce.core.RedisClient
-import io.lettuce.core.RedisFuture
-import io.lettuce.core.RedisURI
 import io.lettuce.core.api.StatefulConnection
-import io.lettuce.core.api.async.RedisAsyncCommands
+import io.lettuce.core.api.reactive.RedisReactiveCommands
 import io.lettuce.core.api.sync.RedisCommands
-import io.lettuce.core.codec.StringCodec
 import io.opentelemetry.auto.test.AgentTestRunner
 import io.opentelemetry.auto.test.utils.PortUtils
+import reactor.core.scheduler.Schedulers
 import redis.embedded.RedisServer
 import spock.lang.Shared
 import spock.util.concurrent.AsyncConditions
 
-import java.util.concurrent.ExecutionException
-import java.util.concurrent.TimeUnit
-import java.util.function.BiConsumer
-import java.util.function.BiFunction
 import java.util.function.Consumer
-import java.util.function.Function
 
+import static io.opentelemetry.auto.test.utils.TraceUtils.runUnderTrace
 import static io.opentelemetry.trace.Span.Kind.CLIENT
 
-class LettuceAsyncClientTest extends AgentTestRunner {
+class LettuceReactiveClientTest extends AgentTestRunner {
   public static final String HOST = "127.0.0.1"
   public static final int DB_INDEX = 0
   // Disable autoreconnect so we do not get stray traces popping up on server shutdown
@@ -48,37 +41,19 @@ class LettuceAsyncClientTest extends AgentTestRunner {
   @Shared
   int port
   @Shared
-  int incorrectPort
-  @Shared
-  String dbAddr
-  @Shared
-  String dbAddrNonExistent
-  @Shared
-  String dbUriNonExistent
-  @Shared
   String embeddedDbUri
 
   @Shared
   RedisServer redisServer
 
-  @Shared
-  Map<String, String> testHashMap = [
-    firstname: "John",
-    lastname : "Doe",
-    age      : "53"
-  ]
-
   RedisClient redisClient
   StatefulConnection connection
-  RedisAsyncCommands<String, ?> asyncCommands
+  RedisReactiveCommands<String, ?> reactiveCommands
   RedisCommands<String, ?> syncCommands
 
   def setupSpec() {
     port = PortUtils.randomOpenPort()
-    incorrectPort = PortUtils.randomOpenPort()
-    dbAddr = HOST + ":" + port + "/" + DB_INDEX
-    dbAddrNonExistent = HOST + ":" + incorrectPort + "/" + DB_INDEX
-    dbUriNonExistent = "redis://" + dbAddrNonExistent
+    String dbAddr = HOST + ":" + port + "/" + DB_INDEX
     embeddedDbUri = "redis://" + dbAddr
 
     redisServer = RedisServer.builder()
@@ -97,7 +72,7 @@ class LettuceAsyncClientTest extends AgentTestRunner {
     redisClient.setOptions(CLIENT_OPTIONS)
 
     connection = redisClient.connect()
-    asyncCommands = connection.async()
+    reactiveCommands = connection.reactive()
     syncCommands = connection.sync()
 
     syncCommands.set("TESTKEY", "TESTVAL")
@@ -109,52 +84,27 @@ class LettuceAsyncClientTest extends AgentTestRunner {
 
   def cleanup() {
     connection.close()
+    redisClient.shutdown()
     redisServer.stop()
   }
 
-  def "connect using get on ConnectionFuture"() {
+  def "set command with subscribe on a defined consumer"() {
     setup:
-    RedisClient testConnectionClient = RedisClient.create(embeddedDbUri)
-    testConnectionClient.setOptions(CLIENT_OPTIONS)
+    def conds = new AsyncConditions()
+    Consumer<String> consumer = new Consumer<String>() {
+      @Override
+      void accept(String res) {
+        conds.evaluate {
+          assert res == "OK"
+        }
+      }
+    }
 
     when:
-    ConnectionFuture connectionFuture = testConnectionClient.connectAsync(StringCodec.UTF8,
-      new RedisURI(HOST, port, 3, TimeUnit.SECONDS))
-    StatefulConnection connection = connectionFuture.get()
+    reactiveCommands.set("TESTSETKEY", "TESTSETVAL").subscribe(consumer)
 
     then:
-    connection != null
-    // Lettuce tracing does not trace connect
-    assertTraces(0) {}
-
-    cleanup:
-    connection.close()
-  }
-
-  def "connect exception inside the connection future"() {
-    setup:
-    RedisClient testConnectionClient = RedisClient.create(dbUriNonExistent)
-    testConnectionClient.setOptions(CLIENT_OPTIONS)
-
-    when:
-    ConnectionFuture connectionFuture = testConnectionClient.connectAsync(StringCodec.UTF8,
-      new RedisURI(HOST, incorrectPort, 3, TimeUnit.SECONDS))
-    StatefulConnection connection = connectionFuture.get()
-
-    then:
-    connection == null
-    thrown ExecutionException
-    // Lettuce tracing does not trace connect
-    assertTraces(0) {}
-  }
-
-  def "set command using Future get with timeout"() {
-    setup:
-    RedisFuture<String> redisFuture = asyncCommands.set("TESTSETKEY", "TESTSETVAL")
-    String res = redisFuture.get(3, TimeUnit.SECONDS)
-
-    expect:
-    res == "OK"
+    conds.await()
     assertTraces(1) {
       trace(0, 1) {
         span(0) {
@@ -180,21 +130,12 @@ class LettuceAsyncClientTest extends AgentTestRunner {
     }
   }
 
-  def "get command chained with thenAccept"() {
+  def "get command with lambda function"() {
     setup:
     def conds = new AsyncConditions()
-    Consumer<String> consumer = new Consumer<String>() {
-      @Override
-      void accept(String res) {
-        conds.evaluate {
-          assert res == "TESTVAL"
-        }
-      }
-    }
 
     when:
-    RedisFuture<String> redisFuture = asyncCommands.get("TESTKEY")
-    redisFuture.thenAccept(consumer)
+    reactiveCommands.get("TESTKEY").subscribe { res -> conds.evaluate { assert res == "TESTVAL" } }
 
     then:
     conds.await()
@@ -225,33 +166,18 @@ class LettuceAsyncClientTest extends AgentTestRunner {
 
   // to make sure instrumentation's chained completion stages won't interfere with user's, while still
   // recording metrics
-  def "get non existent key command with handleAsync and chained with thenApply"() {
+  def "get non existent key command"() {
     setup:
     def conds = new AsyncConditions()
-    final String successStr = "KEY MISSING"
-    BiFunction<String, Throwable, String> firstStage = new BiFunction<String, Throwable, String>() {
-      @Override
-      String apply(String res, Throwable throwable) {
-        conds.evaluate {
-          assert res == null
-          assert throwable == null
-        }
-        return (res == null ? successStr : res)
-      }
-    }
-    Function<String, Object> secondStage = new Function<String, Object>() {
-      @Override
-      Object apply(String input) {
-        conds.evaluate {
-          assert input == successStr
-        }
-        return null
-      }
-    }
+    final defaultVal = "NOT THIS VALUE"
 
     when:
-    RedisFuture<String> redisFuture = asyncCommands.get("NON_EXISTENT_KEY")
-    redisFuture.handleAsync(firstStage).thenApply(secondStage)
+    reactiveCommands.get("NON_EXISTENT_KEY").defaultIfEmpty(defaultVal).subscribe {
+      res ->
+        conds.evaluate {
+          assert res == defaultVal
+        }
+    }
 
     then:
     conds.await()
@@ -278,23 +204,20 @@ class LettuceAsyncClientTest extends AgentTestRunner {
         }
       }
     }
+
   }
 
-  def "command with no arguments using a biconsumer"() {
+  def "command with no arguments"() {
     setup:
     def conds = new AsyncConditions()
-    BiConsumer<String, Throwable> biConsumer = new BiConsumer<String, Throwable>() {
-      @Override
-      void accept(String keyRetrieved, Throwable throwable) {
-        conds.evaluate {
-          assert keyRetrieved != null
-        }
-      }
-    }
 
     when:
-    RedisFuture<String> redisFuture = asyncCommands.randomkey()
-    redisFuture.whenCompleteAsync(biConsumer)
+    reactiveCommands.randomkey().subscribe {
+      res ->
+        conds.evaluate {
+          assert res == "TESTKEY"
+        }
+    }
 
     then:
     conds.await()
@@ -322,46 +245,15 @@ class LettuceAsyncClientTest extends AgentTestRunner {
     }
   }
 
-  def "hash set and then nest apply to hash getall"() {
+  def "command flux publisher "() {
     setup:
-    def conds = new AsyncConditions()
+    reactiveCommands.command().subscribe()
 
-    when:
-    RedisFuture<String> hmsetFuture = asyncCommands.hmset("TESTHM", testHashMap)
-    hmsetFuture.thenApplyAsync(new Function<String, Object>() {
-      @Override
-      Object apply(String setResult) {
-        conds.evaluate {
-          assert setResult == "OK"
-        }
-        RedisFuture<Map<String, String>> hmGetAllFuture = asyncCommands.hgetall("TESTHM")
-        hmGetAllFuture.exceptionally(new Function<Throwable, Map<String, String>>() {
-          @Override
-          Map<String, String> apply(Throwable throwable) {
-            println("unexpected:" + throwable.toString())
-            throwable.printStackTrace()
-            assert false
-            return null
-          }
-        })
-        hmGetAllFuture.thenAccept(new Consumer<Map<String, String>>() {
-          @Override
-          void accept(Map<String, String> hmGetAllResult) {
-            conds.evaluate {
-              assert testHashMap == hmGetAllResult
-            }
-          }
-        })
-        return null
-      }
-    })
-
-    then:
-    conds.await()
-    assertTraces(2) {
+    expect:
+    assertTraces(1) {
       trace(0, 1) {
         span(0) {
-          operationName "HMSET"
+          operationName "COMMAND"
           spanKind CLIENT
           errored false
           tags {
@@ -369,8 +261,8 @@ class LettuceAsyncClientTest extends AgentTestRunner {
             "net.peer.ip" "127.0.0.1"
             "net.peer.port" port
             "db.url" "redis://127.0.0.1:$port"
+            "db.statement" ""
             "db.type" "redis"
-            "db.statement" "key<TESTHM> key<firstname> value<John> key<lastname> value<Doe> key<age> value<53>"
           }
           event(0) {
             eventName "redis.encode.start"
@@ -380,18 +272,193 @@ class LettuceAsyncClientTest extends AgentTestRunner {
           }
         }
       }
-      trace(1, 1) {
+    }
+  }
+
+  def "non reactive command should not produce span"() {
+    setup:
+    String res = null
+
+    when:
+    res = reactiveCommands.digest()
+
+    then:
+    res != null
+    TEST_WRITER.traces.size() == 0
+  }
+
+  def "blocking subscriber"() {
+    when:
+    runUnderTrace("test-parent") {
+      reactiveCommands.set("a", "1")
+        .then(reactiveCommands.get("a"))
+        .block()
+    }
+
+    then:
+    assertTraces(1) {
+      trace(0, 3) {
         span(0) {
-          operationName "HGETALL"
+          operationName "test-parent"
+          errored false
+          tags {
+          }
+        }
+        span(1) {
+          operationName "SET"
           spanKind CLIENT
           errored false
+          childOf span(0)
           tags {
             "net.transport" "IP.TCP"
             "net.peer.ip" "127.0.0.1"
             "net.peer.port" port
             "db.url" "redis://127.0.0.1:$port"
             "db.type" "redis"
-            "db.statement" "key<TESTHM>"
+            "db.statement" "key<a> value<1>"
+          }
+          event(0) {
+            eventName "redis.encode.start"
+          }
+          event(1) {
+            eventName "redis.encode.end"
+          }
+        }
+        span(2) {
+          operationName "GET"
+          spanKind CLIENT
+          errored false
+          childOf span(0)
+          tags {
+            "net.transport" "IP.TCP"
+            "net.peer.ip" "127.0.0.1"
+            "net.peer.port" port
+            "db.url" "redis://127.0.0.1:$port"
+            "db.type" "redis"
+            "db.statement" "key<a>"
+          }
+          event(0) {
+            eventName "redis.encode.start"
+          }
+          event(1) {
+            eventName "redis.encode.end"
+          }
+        }
+      }
+    }
+  }
+
+  def "async subscriber"() {
+    when:
+    runUnderTrace("test-parent") {
+      reactiveCommands.set("a", "1")
+        .then(reactiveCommands.get("a"))
+        .subscribe()
+    }
+
+    then:
+    assertTraces(1) {
+      trace(0, 3) {
+        span(0) {
+          operationName "test-parent"
+          errored false
+          tags {
+          }
+        }
+        span(1) {
+          operationName "SET"
+          spanKind CLIENT
+          errored false
+          childOf span(0)
+          tags {
+            "net.transport" "IP.TCP"
+            "net.peer.ip" "127.0.0.1"
+            "net.peer.port" port
+            "db.url" "redis://127.0.0.1:$port"
+            "db.type" "redis"
+            "db.statement" "key<a> value<1>"
+          }
+          event(0) {
+            eventName "redis.encode.start"
+          }
+          event(1) {
+            eventName "redis.encode.end"
+          }
+        }
+        span(2) {
+          operationName "GET"
+          spanKind CLIENT
+          errored false
+          childOf span(0)
+          tags {
+            "net.transport" "IP.TCP"
+            "net.peer.ip" "127.0.0.1"
+            "net.peer.port" port
+            "db.url" "redis://127.0.0.1:$port"
+            "db.type" "redis"
+            "db.statement" "key<a>"
+          }
+          event(0) {
+            eventName "redis.encode.start"
+          }
+          event(1) {
+            eventName "redis.encode.end"
+          }
+        }
+      }
+    }
+  }
+
+  def "async subscriber with specific thread pool"() {
+    when:
+    runUnderTrace("test-parent") {
+      reactiveCommands.set("a", "1")
+        .then(reactiveCommands.get("a"))
+        .subscribeOn(Schedulers.elastic())
+        .subscribe()
+    }
+
+    then:
+    assertTraces(1) {
+      trace(0, 3) {
+        span(0) {
+          operationName "test-parent"
+          errored false
+          tags {
+          }
+        }
+        span(1) {
+          operationName "SET"
+          spanKind CLIENT
+          errored false
+          childOf span(0)
+          tags {
+            "net.transport" "IP.TCP"
+            "net.peer.ip" "127.0.0.1"
+            "net.peer.port" port
+            "db.url" "redis://127.0.0.1:$port"
+            "db.type" "redis"
+            "db.statement" "key<a> value<1>"
+          }
+          event(0) {
+            eventName "redis.encode.start"
+          }
+          event(1) {
+            eventName "redis.encode.end"
+          }
+        }
+        span(2) {
+          operationName "GET"
+          spanKind CLIENT
+          errored false
+          childOf span(0)
+          tags {
+            "net.transport" "IP.TCP"
+            "net.peer.ip" "127.0.0.1"
+            "net.peer.port" port
+            "db.url" "redis://127.0.0.1:$port"
+            "db.type" "redis"
+            "db.statement" "key<a>"
           }
           event(0) {
             eventName "redis.encode.start"
