+++ settings.gradle
@@ -104,7 +104,7 @@ include ':instrumentation:kafka-streams-0.11'
 include ':instrumentation:khttp-0.1'
 include ':instrumentation:lettuce:lettuce-4.0'
 include ':instrumentation:lettuce:lettuce-5.0'
-include ':instrumentation:lettuce:lettuce-5.2'
+include ':instrumentation:lettuce:lettuce-5.1'
 include ':instrumentation:log4j:log4j-1.1'
 include ':instrumentation:log4j:log4j-2.0'
 include ':instrumentation:logback-1.0'
